#82 Cockpit Sessions tab: embed the reusable session browser
Extract the session browser UI into a reusable `SessionBrowser` component
mounted by both `sandcastle:browse` and the Cockpit's Sessions tab — one
implementation, two mount points, no UI duplication.

- SessionBrowser.tsx: the browser view (tree/detail/pager + loadManifest/
  windowLabelOf), with a `standalone` prop so an embedded instance yields
  q/quit to the Cockpit shell.
- session-browser.tsx: slimmed to the standalone entry (argv → manifest →
  render); `sandcastle:browse` unchanged.
- cockpit.tsx: Sessions tab lazily loads the manifest and mounts
  <SessionBrowser standalone={false}>; input routed through routeCockpitInput.
- cockpit-core.mts: new pure routeCockpitInput — reserves quit + Tab/Shift+Tab
  tab-switch and delegates every other key to the focused tab, resolving the
  ←/→ collision between the shell and the embedded browser. Unit-tested.

Tab-switching moves from ←/→ to Tab/Shift+Tab so ←/→ is free for the browser.
